+++ modules/ui_video.py
@@ -37,9 +37,24 @@ def create_ui():
                     mp4_fps, mp4_interpolate, mp4_codec, mp4_ext, mp4_opt, mp4_video, mp4_frames, mp4_sf, mp4_thumb = video_ui.create_ui_outputs()
                 with gr.Tab('Extras', id='video-extras-tab', elem_id='video_extras') as _video_extras_tab:
                     video_script_inputs = scripts_manager.scripts_video.setup_ui(parent='video', accordion=True)
-                with gr.Tab('Generic', id='video-core-tab') as video_core_tab:
+                with gr.Tab('Models', id='video-core-tab') as video_core_tab:
                     from modules.video_models import video_ui
                     engine, model, steps, sampler_index, width, height, frames, seed = video_ui.create_ui(
+                        prompt, negative, styles,
+                        overrides, video_script_inputs,
+                        mp4_fps, mp4_interpolate, mp4_codec, mp4_ext, mp4_opt, mp4_video, mp4_frames, mp4_sf, mp4_thumb,
+                    )
+                with gr.Tab('MiniMax', id='minimax-tab') as minimax_tab:
+                    from modules.minimax import minimax_ui
+                    minimax_ui.create_ui(
+                        prompt, negative, styles,
+                        overrides, video_script_inputs,
+                        mp4_fps, mp4_interpolate, mp4_codec, mp4_ext, mp4_opt, mp4_video, mp4_frames, mp4_sf, mp4_thumb,
+                    )
+
+                with gr.Tab('LTXVideo', id='ltx-tab') as ltx_tab:
+                    from modules.ltx import ltx_ui
+                    ltx_ui.create_ui(
                         prompt, negative, styles,
                         overrides, video_script_inputs,
                         mp4_fps, mp4_interpolate, mp4_codec, mp4_ext, mp4_opt, mp4_video, mp4_frames, mp4_sf, mp4_thumb,
@@ -51,14 +66,6 @@ def create_ui():
                         overrides, video_script_inputs,
                         mp4_fps, mp4_interpolate, mp4_codec, mp4_ext, mp4_opt, mp4_video, mp4_frames, mp4_sf, mp4_thumb,
                     )
-                with gr.Tab('LTX', id='ltx-tab') as ltx_tab:
-                    from modules.ltx import ltx_ui
-                    ltx_ui.create_ui(
-                        prompt, negative, styles,
-                        overrides, video_script_inputs,
-                        mp4_fps, mp4_interpolate, mp4_codec, mp4_ext, mp4_opt, mp4_video, mp4_frames, mp4_sf, mp4_thumb,
-                    )
-
 
         paste_fields = [
             (prompt, "Prompt"), # cannot add more fields as they are not defined yet
@@ -81,7 +88,8 @@ def create_ui():
 
         current_tab = gr.Textbox(visible=False, value='video')
         video_core_tab.select(fn=lambda: 'video', inputs=[], outputs=[current_tab])
-        framepack_tab.select(fn=lambda: 'framepack', inputs=[], outputs=[current_tab])
+        minimax_tab.select(fn=lambda: 'minimax', inputs=[], outputs=[current_tab])
         ltx_tab.select(fn=lambda: 'ltx', inputs=[], outputs=[current_tab])
+        framepack_tab.select(fn=lambda: 'framepack', inputs=[], outputs=[current_tab])
 
         generate_btn.click(fn=None, _js='submit_video_wrapper', inputs=[current_tab], outputs=[])